Prepared Enemy Animations, improved general code

+++ God Damn Hell/Assets/Scripts/Player/PlayerStats.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public class PlayerStats : MonoBehaviour
 {
-    public float healthPoints;
+    public int healthPoints;
 }
